+++ docs/CHANGELOG.md
@@ -109,6 +109,7 @@
 - fixed Puna's summoned Lizards counting towards total level kill count
 - fixed Tony briefly appearing for a single frame when loading a save after his death
 - fixed Lara sometimes getting stuck when crawling backwards off a tilted ledge (#4956)
+- fixed the Tribe Pipeman sometimes not being able to aim darts at Lara correctly (Gameplay → Fixes → Fix Pipeman aim)
 - removed the limitation of one Carcass instance per level working with Piranhas
 - removed the limitation of Piranhas only attacking Carcass instances if the level sequence matches Crash Site's
 - restored the ability for Lara to perform grab cancels, like TR1 and TR2